+++ ContC.domain.entities/Models/Receita.cs
@@ -14,7 +14,7 @@ namespace ContC.domain.entities.Models
 
         public virtual int Id { get; set; }
         public virtual string Descricao { get; set; }
-        public virtual double Valor { get; set; }
+        public virtual decimal Valor { get; set; }
         public virtual DateTime DataCadastro { get; set; }
         public virtual DateTime DataRecebimento { get; set; }
         public virtual Grupo Endereco { get; set; }
